Using the subjects from the client certs for auth
Since the certificates to access NATs can be created by the user in any
way they want, and these certificates are for the Director and the
Health Monitor, and this components will always require access to NATs,
we need no make sure that the subjects on those certificates have
access and the right permissions on NATs.
To be sure that we are granting access, we are now getting those
certificates on the NATs job, and taking from them the subject and
adding that subject to the NATs auth file.